pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "explerr"
version = "0.1.0"
description = "Error explainer using LLMs"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Honghua Dong", email = "dhh19951@gmail.com" }
]
dependencies = [
    "applang>=0.1.5",
    "rich",
]

[project.scripts]
expython = "explerr.cli:main"
